requirements.txt
pymediainfo==2.1.9
requests==2.18.1
Pillow==4.3.0
PyQt4==4.11.4
beautifulsoup4==4.6.0
